+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file documents the changes made to the formatter with each release.
 
 - Make sure to keep tool at the top of the script, above class_name and extends (#285)
 - Fixed generic type parameters breaking across lines (#283)
+- Fixed is not being lost in a parenthesized expression (#284)
 
 ## Release 0.22.0